Fix: 다이어리 삭제 복수개 가능하도록 변경

+++ src/main/java/what/what2eat/domain/diary/service/DiaryService.java
@@ -27,6 +27,7 @@ import java.io.IOException;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 @Service
 @Slf4j
@@ -77,7 +78,7 @@ public class DiaryService {
         existingDiary.update(existingDiary);
 
         // 기존 이미지 URL 리스트
-        List<String> existingUrlList = existingDiary.getUploadImgList().stream()
+        List<String> existingUrlList = existingDiary.getDiaryImageList().stream()
                 .map(diaryImage -> diaryImage.getImageUrl())
                 .collect(Collectors.toList());
 
@@ -93,7 +94,7 @@ public class DiaryService {
             s3Service.deleteFiles(deleteUrlList);
 
             // DB 데이터 삭제
-            existingDiary.getUploadImgList().removeIf(
+            existingDiary.getDiaryImageList().removeIf(
                     diaryImage -> deleteUrlList.contains(diaryImage.getImageUrl())
             );
         }
@@ -109,21 +110,26 @@ public class DiaryService {
     }
 
     // 다이어리 삭제
-    public void deleteDiary(Long diaryId) {
-        // 다이어리 조회
-        Diary existingDiary = diaryRepository.findById(diaryId)
-                .orElseThrow(() -> new DiaryException(DiaryErrorCode.DIARY_NOT_FOUND));
+    public void deleteDiary(DiaryRequestDTO.DiaryDeleteDTO request) {
+
+        // 다이어리 목록 조회
+        List<Diary> existingDiaryList = diaryRepository.findAllById(request.getDiaryIdList());
+
+        if (existingDiaryList.isEmpty()) {
+            throw new DiaryException(DiaryErrorCode.DIARY_NOT_FOUND);
+        }
 
         // s3에 저장된 imgUrl List 반환
-        List<String> uploadedImgUrlList = existingDiary.getUploadImgList().stream()
-                .map(diaryImage -> diaryImage.getImageUrl())
+        List<String> uploadedImgUrlList = existingDiaryList.stream()
+                .flatMap(diary -> diary.getDiaryImageList().stream())
+                .map(DiaryImage::getImageUrl)
                 .collect(Collectors.toList());
 
         // s3에 저장된 파일 삭제
         s3Service.deleteFiles(uploadedImgUrlList);
 
         // DB에 저장된 다이어리, 다이어리 이미지 삭제
-        diaryRepository.delete(existingDiary);
+        diaryRepository.deleteAll(existingDiaryList);
 
     }
 
@@ -170,7 +176,7 @@ public class DiaryService {
                 .collect(Collectors.toList());
 
         // Diary 엔티티에 연관관계 설정
-        diary.getUploadImgList().addAll(imageList);
+        diary.getDiaryImageList().addAll(imageList);
     }
 
     // 위도, 경도 -> point로 변환
